+++ tergite_autocalibration/tools/cli/__init__.py
@@ -11,188 +11,55 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-from pathlib import Path
-
-import click
 import typer
 
-from tergite_autocalibration.utils.dto.enums import MeasurementMode
+from .calibration import calibration_cli
 from .cluster import cluster_cli
 from .graph import graph_cli
-from .joke import joke_cli
+from .node import node_cli
 
-app = typer.Typer(no_args_is_help=True)
-app.add_typer(cluster_cli,
-              name="cluster",
-              help="Handle operations related to the cluster.",
-              no_args_is_help=True)
-app.add_typer(joke_cli,
-              name="joke",
-              help="Tells a joke.",
-              no_args_is_help=True)
-app.add_typer(graph_cli,
-              name="graph",
-              help="Handle operations related to the calibration graph.",
-              no_args_is_help=True)
+cli = typer.Typer(no_args_is_help=True)
 
-
-
-@click.group()
-def cli():
-    pass
-
-
-@cli.group(help="Handle operations related to the node.")
-def node():
-    pass
-
-
-@node.command(help="Reset all parameters in redis for the node specified.")
-@click.option(
-    "-n",
-    "--name",
-    required=False,
-    help="Name of the node to be reset in redis e.g resonator_spectroscopy.",
+cli.add_typer(
+    calibration_cli,
+    name="calibration",
+    help="Start the calibration supervisor.",
+    no_args_is_help=True,
 )
-@click.option(
-    "-a",
-    "--all",
-    required=False,
-    is_flag=True,
-    help="Use -a if you want to reset all nodes.",
+cli.add_typer(
+    node_cli,
+    name="node",
+    help="Handle operations related to the node.",
+    no_args_is_help=True,
 )
-@click.option(
-    "-f",
-    "--from_node",
-    required=False,
-    help="Use -f node_name if you want to reset all nodes from specified node in chain.",
+cli.add_typer(
+    cluster_cli,
+    name="cluster",
+    help="Handle operations related to the cluster.",
+    no_args_is_help=True,
 )
-def reset(name, all, from_node):
-    from tergite_autocalibration.utils.reset_redis_node import ResetRedisNode
-    from tergite_autocalibration.lib.utils.graph import range_topological_order
-    from tergite_autocalibration.utils.user_input import user_requested_calibration
+cli.add_typer(
+    graph_cli,
+    name="graph",
+    help="Handle operations related to the calibration graph.",
+    no_args_is_help=True,
+)
 
-    topo_order = range_topological_order(
-        from_node, user_requested_calibration["target_node"]
-    )
 
-    reset_obj_ = ResetRedisNode()
-    if from_node:
-        if click.confirm(
-                "Do you really want to reset all nodes from"
-                + from_node
-                + "? It might take some time to recalibrate them."
-        ):
-            for node in topo_order:
-                reset_obj_.reset_node(node)
+@cli.command(help="Tell a joke.")
+def joke():
+    import asyncio
+    from jokeapi import Jokes
+
+    async def print_joke():
+        j = await Jokes()  # Initialise the class
+        joke_ = await j.get_joke(
+            blacklist=["racist", "religious", "political", "nsfw", "sexist"]
+        )  # Retrieve a random joke
+        if joke_["type"] == "single":  # Print the joke
+            print(joke_["joke"])
         else:
-            click.echo("Node reset aborted by user.")
-    elif all:
-        if click.confirm(
-                "Do you really want to reset all nodes? It might take some time to recalibrate them."
-        ):
-            reset_obj_.reset_node("all")
-        else:
-            click.echo("Node reset aborted by user.")
-    elif name is not None:
-        reset_obj_.reset_node(name)
-    else:
-        click.echo("Please enter a node name or use the -a option to reset all nodes.")
+            print(joke_["setup"])
+            print(joke_["delivery"])
 
-
-@cli.group(help="Handle operations related to the calibration supervisor.")
-def calibration():
-    pass
-
-
-@calibration.command(help="Start the calibration supervisor.")
-@click.option(
-    "-c",
-    required=False,
-    help="Takes the cluster ip address as argument. If not set, it will try take CLUSTER_IP in the .env file.",
-)
-@click.option(
-    "-d",
-    required=False,
-    is_flag=True,
-    help="Use -d if you want to use the dummy cluster (not implemented)",
-)
-@click.option(
-    "-r",
-    required=False,
-    help="Use -r if you want to use rerun an analysis, give the path to the dataset folder (plots will be overwritten), you also need to specify the name of the node using -n",
-)
-@click.option(
-    "-n",
-    "--name",
-    required=False,
-    help="Use to specify the node type to rerun, only works with -r option",
-)
-@click.option(
-    "--push",
-    required=False,
-    is_flag=True,
-    help="If --push the a backend will pushed to an MSS specified in MSS_MACHINE_ROOT_URL in the .env file.",
-)
-def start(c, d, r, name, push):
-    from ipaddress import ip_address, IPv4Address
-
-    from tergite_autocalibration.config.settings import CLUSTER_IP
-    from tergite_autocalibration.scripts.calibration_supervisor import (
-        CalibrationSupervisor,
-    )
-    from tergite_autocalibration.scripts.db_backend_update import update_mss
-
-    cluster_mode: "MeasurementMode" = MeasurementMode.real
-    parsed_cluster_ip: "IPv4Address" = CLUSTER_IP
-    node_name = ""
-    data_path = ""
-
-    # Checks whether to start the cluster in dummy mode
-    # TODO: The dummy cluster is currently not implemented
-    if d:
-        click.echo(
-            "The option to run on a dummy cluster is currently not implemented. "
-            "Trying to start the calibration supervisor with default cluster configuration"
-        )
-        cluster_mode = MeasurementMode.dummy
-
-    if r:
-        folder_path = Path(r)
-
-        # Check if the folder exists
-        if not folder_path.is_dir():
-            print(f"Error: The specified folder '{folder_path}' does not exist.")
-            exit(1)  # Exit with an error code
-
-        if not name:
-            click.echo(
-                "You are trying to re-run the analysis on a specific node but you did not specify it."
-                "Please specify the node using -n or --name."
-            )
-            exit(1)  # Exit with an error exit
-
-        cluster_mode = MeasurementMode.re_analyse
-        data_path = folder_path
-        node_name = name
-
-    # Check whether the ip address of the cluster is set correctly
-    if c and not d:
-        if len(c) >= 0:
-            cluster_mode = MeasurementMode.real
-            parsed_cluster_ip = ip_address(c)
-        else:
-            click.echo(
-                "Cluster argument requires the ip address of the cluster as parameter. "
-                "Trying to start the calibration supervisor with default cluster configuration."
-            )
-
-    supervisor = CalibrationSupervisor(
-        cluster_mode=cluster_mode,
-        cluster_ip=parsed_cluster_ip,
-        node_name=node_name,
-        data_path=data_path,
-    )
-    supervisor.calibrate_system()
-    if push:
-        update_mss()
+    asyncio.run(print_joke())